@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 class CustomOnbaordingButton extends StatelessWidget {
-  final GestureTapCallback onPressed;
+  final void Function()? onPressed;
   final String title;
 
   const CustomOnbaordingButton({
@@ -17,11 +17,39 @@ class CustomOnbaordingButton extends StatelessWidget {
     return Container(
       width: double.infinity,
       margin: const EdgeInsets.symmetric(horizontal: 20),
-      child: MaterialButton(
-        padding: const EdgeInsets.symmetric(vertical: 18),
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(12),
-          side: const BorderSide(color: Colors.pink),
+      child: ElevatedButton(
+        style: ButtonStyle(
+          padding: MaterialStateProperty.all<EdgeInsets>(
+            const EdgeInsets.symmetric(vertical: 18),
+          ),
+          shape: MaterialStateProperty.resolveWith<OutlinedBorder>(
+            (Set<MaterialState> states) {
+              if (states.contains(MaterialState.pressed)) {
+                return RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(12),
+                  side: const BorderSide(color: Colors.pink),
+                );
+              } else if (states.contains(MaterialState.disabled)) {
+                return RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(12),
+                );
+              }
+              return RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(12),
+                side: const BorderSide(color: Colors.pink),
+              );
+            },
+          ),
+          backgroundColor: MaterialStateProperty.resolveWith<Color>(
+            (Set<MaterialState> states) {
+              if (states.contains(MaterialState.pressed)) {
+                return ToDoTheme.primaryColor;
+              } else if (states.contains(MaterialState.disabled)) {
+                return ToDoTheme.textFieldColor;
+              }
+              return ToDoTheme.primaryColor;
+            },
+          ),
         ),
         child: Text(
           title,
@@ -31,7 +59,6 @@ class CustomOnbaordingButton extends StatelessWidget {
             fontWeight: FontWeight.w600,
           ),
         ),
-        color: AppCommonTheme.primaryColor,
         onPressed: onPressed,
       ),
     );
