@@ -327,7 +327,7 @@ class NotificationPopUpTheme {
     TextStyle(
       fontSize: 15,
       fontWeight: FontWeight.w300,
-      color : networkTextColor,
+      color: networkTextColor,
     ),
   );
 }
@@ -360,6 +360,7 @@ class ToDoTheme {
   static const subscribedIconColor = ColorRef(Color(0xFFFCE5E5));
   static const taskIconColor = ColorRef(Color(0xFFEAD0FF));
   static const todayCalendarColor = ColorRef(Color(0xFF23AFC2));
+  static const textFieldColor = ColorRef(Color(0xE6333540));
   static const unAssignedIconColor = ColorRef(Color(0xFFFFE9C9));
 
   // Text Scheme.
@@ -369,8 +370,8 @@ class ToDoTheme {
     fontWeight: FontWeight.w700,
   );
 
-  static const subtitleTextStyle = TextStyle(
-    color: primaryTextColor,
+  static const descriptionTextStyle = TextStyle(
+    color: calendarColor,
     fontSize: 14,
     fontWeight: FontWeight.w500,
   );
@@ -382,7 +383,7 @@ class ToDoTheme {
   );
 
   static const listSubtitleTextStyle = TextStyle(
-    color: secondaryTextColor,
+    color: calendarColor,
     fontSize: 14,
     fontWeight: FontWeight.w400,
   );
@@ -470,6 +471,18 @@ class ToDoTheme {
     fontWeight: FontWeight.w500,
   );
 
+  static const textFieldCounterStyle = TextStyle(
+    color: secondaryTextColor,
+    fontSize: 12,
+    fontWeight: FontWeight.w500,
+  );
+
+  static const selectTeamTextStyle = TextStyle(
+    color: primaryTextColor,
+    fontSize: 13,
+    fontWeight: FontWeight.w400,
+  );
+
   // Background Scheme.
   static const toDoDecoration = BoxDecoration(
     gradient: LinearGradient(
@@ -488,6 +501,7 @@ class PinsTheme {
   static const backgroundGradientColor = ColorRef(Color(0xFF242632));
   static const backgroundGradient2Color = ColorRef(Color(0x885C2A80));
   static const primaryTextColor = Color.fromARGB(255, 255, 255, 255);
+  static const cardBackgroundColor = ColorRef(Color(0xff2F313E));
 
   // Text Scheme.
   // Background Scheme.
