@@ -1,6 +1,7 @@
 import 'package:effektio/common/store/themes/SeperatedThemes.dart';
 import 'package:effektio/controllers/FaqController.dart';
 import 'package:effektio/widgets/FaqListItem.dart';
+import 'package:effektio/widgets/searchWidget.dart';
 import 'package:effektio_flutter_sdk/effektio_flutter_sdk_ffi.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
@@ -38,100 +39,56 @@ class _FaqOverviewScreenState extends State<FaqOverviewScreen> {
                 ),
               )
             : Scaffold(
-                body: Container(
-                  decoration: PinsTheme.pinsDecoration,
-                  child: Padding(
-                    padding: const EdgeInsets.only(top: 25),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: <Widget>[
-                        const Padding(
-                          padding: EdgeInsets.only(left: 12),
-                          child: Text(
-                            'Pins',
-                            style: PinsTheme.titleTextStyle,
-                          ),
+                body: Padding(
+                  padding: const EdgeInsets.only(top: 25),
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: <Widget>[
+                      const Padding(
+                        padding: EdgeInsets.only(left: 12),
+                        child: Text(
+                          'Pins',
+                          style: PinsTheme.titleTextStyle,
                         ),
-                        Padding(
-                          padding: const EdgeInsets.only(
-                            top: 5,
-                            bottom: 6,
-                            left: 10,
-                            right: 10,
-                          ),
-                          child: TextField(
-                            onChanged: (value) {
-                              faqController.searchedData(
-                                value,
-                                snapshot,
-                              );
-                            },
-                            controller: faqController.searchController,
-                            style: ToDoTheme.taskTitleTextStyle.copyWith(
-                              fontWeight: FontWeight.w500,
+                      ),
+                      SearchWidget(
+                        searchController: faqController.searchController,
+                        onChanged: (text) {
+                          faqController.searchedData(
+                            text.toString(),
+                            snapshot,
+                          );
+                        },
+                        onReset: () {
+                          faqController.searchData.clear();
+                          setState(() {});
+                        },
+                      ),
+                      GetBuilder<FaqController>(
+                        builder: (FaqController controller) {
+                          return Expanded(
+                            child: ListView.builder(
+                              padding: const EdgeInsets.all(8),
+                              shrinkWrap: true,
+                              itemCount: controller.searchController.text.isEmpty
+                                  ? snapshot.requireData.length
+                                  : controller.searchData.length,
+                              itemBuilder: (
+                                BuildContext context,
+                                int index,
+                              ) {
+                                return FaqListItem(
+                                  client: widget.client,
+                                  faq: controller.searchData.isEmpty
+                                      ? snapshot.requireData[index]
+                                      : controller.searchData[index],
+                                );
+                              },
                             ),
-                            cursorColor: ToDoTheme.primaryTextColor,
-                            decoration: InputDecoration(
-                              hintStyle: const TextStyle(
-                                color: Colors.white,
-                                fontSize: 13,
-                              ),
-                              suffixIcon: const Icon(
-                                Icons.search,
-                                color: Colors.white,
-                              ),
-                              contentPadding: const EdgeInsets.only(
-                                left: 12,
-                                bottom: 2,
-                                top: 2,
-                              ),
-                              border: OutlineInputBorder(
-                                borderSide: const BorderSide(
-                                  color: Colors.white,
-                                ),
-                                borderRadius: BorderRadius.circular(20.0),
-                              ),
-                              focusedBorder: OutlineInputBorder(
-                                borderSide: const BorderSide(
-                                  color: Colors.white,
-                                ),
-                                borderRadius: BorderRadius.circular(30.0),
-                              ),
-                              enabledBorder: OutlineInputBorder(
-                                borderSide: const BorderSide(
-                                  color: Colors.white,
-                                ),
-                                borderRadius: BorderRadius.circular(30.0),
-                              ),
-                            ),
-                          ),
-                        ),
-                        GetBuilder<FaqController>(
-                          builder: (FaqController controller) {
-                            return Expanded(
-                              child: ListView.builder(
-                                padding: const EdgeInsets.all(8),
-                                shrinkWrap: true,
-                                itemCount: controller.searchData.isEmpty
-                                    ? snapshot.requireData.length
-                                    : controller.searchData.length,
-                                itemBuilder: (
-                                  BuildContext context,
-                                  int index,
-                                ) {
-                                  return FaqListItem(
-                                    client: widget.client,
-                                    faq: controller.searchData.isEmpty
-                                        ? snapshot.requireData[index]
-                                        : controller.searchData[index],
-                                  );
-                                },
-                              ),
-                            );
-                          },
-                        ),
-                      ],
-                    ),
+                          );
+                        },
+                      ),
+                    ],
                   ),
                 ),
               );
