@@ -1,3 +1,4 @@
+import 'package:duration/duration.dart';
 import 'package:effektio/common/animations/LikeAnimation.dart';
 import 'package:effektio/common/store/themes/SeperatedThemes.dart';
 import 'package:effektio/widgets/NewsItem.dart';
@@ -31,13 +32,12 @@ class _NewsScreenState extends State<NewsScreen>
 
     controller = AnimationController(
       vsync: this,
-      duration: const Duration(seconds: 1),
+      duration: seconds(1),
     );
   }
 
   @override
   Widget build(BuildContext context) {
-
     return FutureBuilder<FfiListNews>(
       future: widget.client.latestNews(),
       builder: (BuildContext context, AsyncSnapshot<FfiListNews> snapshot) {
@@ -142,6 +142,7 @@ class _NewsScreenState extends State<NewsScreen>
               userId: widget.client.userId().toString(),
               displayName: widget.displayName,
               displayAvatar: widget.displayAvatar,
+              client: widget.client,
             ),
             body: PageView.builder(
               itemCount: snapshot.requireData.length,
