@@ -104,7 +104,7 @@ class ToDoTaskAssignScreen extends StatelessWidget {
                       ),
                       Text(
                         '${avatars.length}',
-                        style: ToDoTheme.subtitleTextStyle,
+                        style: ToDoTheme.descriptionTextStyle,
                       ),
                     ],
                   ),
@@ -141,14 +141,15 @@ class ToDoTaskAssignScreen extends StatelessWidget {
               crossAxisAlignment: CrossAxisAlignment.start,
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                Text('Someone not in the list? Just Invite them',
+                Text(
+                  'Someone not in the list? Just Invite them',
                   style: ToDoTheme.taskTitleTextStyle.copyWith(
                     fontWeight: FontWeight.w500,
                     fontSize: 16,
                   ),
                 ),
                 InkWell(
-                  onTap: (){
+                  onTap: () {
                     showDialogBox(context);
                   },
                   child: Container(
@@ -161,11 +162,15 @@ class ToDoTaskAssignScreen extends StatelessWidget {
                     child: Row(
                       mainAxisAlignment: MainAxisAlignment.center,
                       children: [
-                        const Icon(Icons.group, color: ToDoTheme.primaryTextColor,),
+                        const Icon(
+                          Icons.group,
+                          color: ToDoTheme.primaryTextColor,
+                        ),
                         const SizedBox(
                           width: 8.0,
                         ),
-                        Text('Invite Members',
+                        Text(
+                          'Invite Members',
                           style: ToDoTheme.taskTitleTextStyle.copyWith(
                             fontWeight: FontWeight.w500,
                             fontSize: 14,
@@ -191,6 +196,4 @@ class ToDoTaskAssignScreen extends StatelessWidget {
       },
     );
   }
-
-
 }
