@@ -868,6 +868,50 @@ class Api {
     return tmp1;
   }
 
+  EfkDuration durationFromSecs(
+    int secs,
+  ) {
+    final tmp0 = secs;
+    var tmp1 = 0;
+    tmp1 = tmp0;
+    final tmp2 = _durationFromSecs(
+      tmp1,
+    );
+    final tmp4 = tmp2;
+    final ffi.Pointer<ffi.Void> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+    final tmp4_1 = _Box(this, tmp4_0, "drop_box_EfkDuration");
+    tmp4_1._finalizer = this._registerFinalizer(tmp4_1);
+    final tmp3 = EfkDuration._(this, tmp4_1);
+    return tmp3;
+  }
+
+  CreateGroupSettings newGroupSettings(
+    String name,
+  ) {
+    final tmp0 = name;
+    var tmp1 = 0;
+    var tmp2 = 0;
+    var tmp3 = 0;
+    final tmp0_0 = utf8.encode(tmp0);
+    tmp2 = tmp0_0.length;
+    final ffi.Pointer<ffi.Uint8> tmp1_0 = this.__allocate(tmp2 * 1, 1);
+    final Uint8List tmp1_1 = tmp1_0.asTypedList(tmp2);
+    tmp1_1.setAll(0, tmp0_0);
+    tmp1 = tmp1_0.address;
+    tmp3 = tmp2;
+    final tmp4 = _newGroupSettings(
+      tmp1,
+      tmp2,
+      tmp3,
+    );
+    final tmp6 = tmp4;
+    final ffi.Pointer<ffi.Void> tmp6_0 = ffi.Pointer.fromAddress(tmp6);
+    final tmp6_1 = _Box(this, tmp6_0, "drop_box_CreateGroupSettings");
+    tmp6_1._finalizer = this._registerFinalizer(tmp6_1);
+    final tmp5 = CreateGroupSettings._(this, tmp6_1);
+    return tmp5;
+  }
+
   late final _allocatePtr = _lookup<
       ffi.NativeFunction<
           ffi.Pointer<ffi.Uint8> Function(ffi.IntPtr, ffi.IntPtr)>>("allocate");
@@ -3905,6 +3949,186 @@ class Api {
     return tmp7;
   }
 
+  RoomId? __clientCreateEffektioGroupFuturePoll(
+    int boxed,
+    int postCobject,
+    int port,
+  ) {
+    final tmp0 = boxed;
+    final tmp2 = postCobject;
+    final tmp4 = port;
+    var tmp1 = 0;
+    var tmp3 = 0;
+    var tmp5 = 0;
+    tmp1 = tmp0;
+    tmp3 = tmp2;
+    tmp5 = tmp4;
+    final tmp6 = _clientCreateEffektioGroupFuturePoll(
+      tmp1,
+      tmp3,
+      tmp5,
+    );
+    final tmp8 = tmp6.arg0;
+    final tmp9 = tmp6.arg1;
+    final tmp10 = tmp6.arg2;
+    final tmp11 = tmp6.arg3;
+    final tmp12 = tmp6.arg4;
+    final tmp13 = tmp6.arg5;
+    if (tmp8 == 0) {
+      return null;
+    }
+    if (tmp9 == 0) {
+      final ffi.Pointer<ffi.Uint8> tmp10_0 = ffi.Pointer.fromAddress(tmp10);
+      final tmp9_0 = utf8.decode(tmp10_0.asTypedList(tmp11));
+      if (tmp11 > 0) {
+        final ffi.Pointer<ffi.Void> tmp10_0;
+        tmp10_0 = ffi.Pointer.fromAddress(tmp10);
+        this.__deallocate(tmp10_0, tmp12, 1);
+      }
+      throw tmp9_0;
+    }
+    final ffi.Pointer<ffi.Void> tmp13_0 = ffi.Pointer.fromAddress(tmp13);
+    final tmp13_1 = _Box(this, tmp13_0, "drop_box_RoomId");
+    tmp13_1._finalizer = this._registerFinalizer(tmp13_1);
+    final tmp7 = RoomId._(this, tmp13_1);
+    return tmp7;
+  }
+
+  Comment? __clientWaitForCommentFuturePoll(
+    int boxed,
+    int postCobject,
+    int port,
+  ) {
+    final tmp0 = boxed;
+    final tmp2 = postCobject;
+    final tmp4 = port;
+    var tmp1 = 0;
+    var tmp3 = 0;
+    var tmp5 = 0;
+    tmp1 = tmp0;
+    tmp3 = tmp2;
+    tmp5 = tmp4;
+    final tmp6 = _clientWaitForCommentFuturePoll(
+      tmp1,
+      tmp3,
+      tmp5,
+    );
+    final tmp8 = tmp6.arg0;
+    final tmp9 = tmp6.arg1;
+    final tmp10 = tmp6.arg2;
+    final tmp11 = tmp6.arg3;
+    final tmp12 = tmp6.arg4;
+    final tmp13 = tmp6.arg5;
+    if (tmp8 == 0) {
+      return null;
+    }
+    if (tmp9 == 0) {
+      final ffi.Pointer<ffi.Uint8> tmp10_0 = ffi.Pointer.fromAddress(tmp10);
+      final tmp9_0 = utf8.decode(tmp10_0.asTypedList(tmp11));
+      if (tmp11 > 0) {
+        final ffi.Pointer<ffi.Void> tmp10_0;
+        tmp10_0 = ffi.Pointer.fromAddress(tmp10);
+        this.__deallocate(tmp10_0, tmp12, 1);
+      }
+      throw tmp9_0;
+    }
+    final ffi.Pointer<ffi.Void> tmp13_0 = ffi.Pointer.fromAddress(tmp13);
+    final tmp13_1 = _Box(this, tmp13_0, "drop_box_Comment");
+    tmp13_1._finalizer = this._registerFinalizer(tmp13_1);
+    final tmp7 = Comment._(this, tmp13_1);
+    return tmp7;
+  }
+
+  TaskList? __clientWaitForTaskListFuturePoll(
+    int boxed,
+    int postCobject,
+    int port,
+  ) {
+    final tmp0 = boxed;
+    final tmp2 = postCobject;
+    final tmp4 = port;
+    var tmp1 = 0;
+    var tmp3 = 0;
+    var tmp5 = 0;
+    tmp1 = tmp0;
+    tmp3 = tmp2;
+    tmp5 = tmp4;
+    final tmp6 = _clientWaitForTaskListFuturePoll(
+      tmp1,
+      tmp3,
+      tmp5,
+    );
+    final tmp8 = tmp6.arg0;
+    final tmp9 = tmp6.arg1;
+    final tmp10 = tmp6.arg2;
+    final tmp11 = tmp6.arg3;
+    final tmp12 = tmp6.arg4;
+    final tmp13 = tmp6.arg5;
+    if (tmp8 == 0) {
+      return null;
+    }
+    if (tmp9 == 0) {
+      final ffi.Pointer<ffi.Uint8> tmp10_0 = ffi.Pointer.fromAddress(tmp10);
+      final tmp9_0 = utf8.decode(tmp10_0.asTypedList(tmp11));
+      if (tmp11 > 0) {
+        final ffi.Pointer<ffi.Void> tmp10_0;
+        tmp10_0 = ffi.Pointer.fromAddress(tmp10);
+        this.__deallocate(tmp10_0, tmp12, 1);
+      }
+      throw tmp9_0;
+    }
+    final ffi.Pointer<ffi.Void> tmp13_0 = ffi.Pointer.fromAddress(tmp13);
+    final tmp13_1 = _Box(this, tmp13_0, "drop_box_TaskList");
+    tmp13_1._finalizer = this._registerFinalizer(tmp13_1);
+    final tmp7 = TaskList._(this, tmp13_1);
+    return tmp7;
+  }
+
+  Task? __clientWaitForTaskFuturePoll(
+    int boxed,
+    int postCobject,
+    int port,
+  ) {
+    final tmp0 = boxed;
+    final tmp2 = postCobject;
+    final tmp4 = port;
+    var tmp1 = 0;
+    var tmp3 = 0;
+    var tmp5 = 0;
+    tmp1 = tmp0;
+    tmp3 = tmp2;
+    tmp5 = tmp4;
+    final tmp6 = _clientWaitForTaskFuturePoll(
+      tmp1,
+      tmp3,
+      tmp5,
+    );
+    final tmp8 = tmp6.arg0;
+    final tmp9 = tmp6.arg1;
+    final tmp10 = tmp6.arg2;
+    final tmp11 = tmp6.arg3;
+    final tmp12 = tmp6.arg4;
+    final tmp13 = tmp6.arg5;
+    if (tmp8 == 0) {
+      return null;
+    }
+    if (tmp9 == 0) {
+      final ffi.Pointer<ffi.Uint8> tmp10_0 = ffi.Pointer.fromAddress(tmp10);
+      final tmp9_0 = utf8.decode(tmp10_0.asTypedList(tmp11));
+      if (tmp11 > 0) {
+        final ffi.Pointer<ffi.Void> tmp10_0;
+        tmp10_0 = ffi.Pointer.fromAddress(tmp10);
+        this.__deallocate(tmp10_0, tmp12, 1);
+      }
+      throw tmp9_0;
+    }
+    final ffi.Pointer<ffi.Void> tmp13_0 = ffi.Pointer.fromAddress(tmp13);
+    final tmp13_1 = _Box(this, tmp13_0, "drop_box_Task");
+    tmp13_1._finalizer = this._registerFinalizer(tmp13_1);
+    final tmp7 = Task._(this, tmp13_1);
+    return tmp7;
+  }
+
   FfiBufferUint8? __userProfileGetAvatarFuturePoll(
     int boxed,
     int postCobject,
@@ -5318,6 +5542,39 @@ class Api {
     return tmp9;
   }
 
+  bool? __clientSubscribeStreamPoll(
+    int boxed,
+    int postCobject,
+    int port,
+    int done,
+  ) {
+    final tmp0 = boxed;
+    final tmp2 = postCobject;
+    final tmp4 = port;
+    final tmp6 = done;
+    var tmp1 = 0;
+    var tmp3 = 0;
+    var tmp5 = 0;
+    var tmp7 = 0;
+    tmp1 = tmp0;
+    tmp3 = tmp2;
+    tmp5 = tmp4;
+    tmp7 = tmp6;
+    final tmp8 = _clientSubscribeStreamPoll(
+      tmp1,
+      tmp3,
+      tmp5,
+      tmp7,
+    );
+    final tmp10 = tmp8.arg0;
+    final tmp11 = tmp8.arg1;
+    if (tmp10 == 0) {
+      return null;
+    }
+    final tmp9 = tmp11 > 0;
+    return tmp9;
+  }
+
   late final _initLoggingPtr = _lookup<
       ffi.NativeFunction<
           _InitLoggingReturn Function(
@@ -5461,14 +5718,38 @@ class Api {
       _lookup<ffi.NativeFunction<ffi.Int64 Function()>>("__gen_mock_news");
 
   late final _genMockNews = _genMockNewsPtr.asFunction<int Function()>();
-  late final _colorRgbaU8Ptr = _lookup<
+  late final _durationFromSecsPtr = _lookup<
       ffi.NativeFunction<
-          _ColorRgbaU8Return Function(
-    ffi.Int64,
-  )>>("__Color_rgba_u8");
+          ffi.Int64 Function(
+    ffi.Uint64,
+  )>>("__duration_from_secs");
 
-  late final _colorRgbaU8 = _colorRgbaU8Ptr.asFunction<
-      _ColorRgbaU8Return Function(
+  late final _durationFromSecs = _durationFromSecsPtr.asFunction<
+      int Function(
+    int,
+  )>();
+  late final _newGroupSettingsPtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Int64 Function(
+    ffi.Int64,
+    ffi.Uint64,
+    ffi.Uint64,
+  )>>("__new_group_settings");
+
+  late final _newGroupSettings = _newGroupSettingsPtr.asFunction<
+      int Function(
+    int,
+    int,
+    int,
+  )>();
+  late final _efkColorRgbaU8Ptr = _lookup<
+      ffi.NativeFunction<
+          _EfkColorRgbaU8Return Function(
+    ffi.Int64,
+  )>>("__EfkColor_rgba_u8");
+
+  late final _efkColorRgbaU8 = _efkColorRgbaU8Ptr.asFunction<
+      _EfkColorRgbaU8Return Function(
     int,
   )>();
   late final _utcDateTimeTimestampPtr = _lookup<
@@ -5742,15 +6023,14 @@ class Api {
           int Function(
     int,
   )>();
-  late final _roomEventItemItemContentTypePtr = _lookup<
+  late final _roomEventItemEventTypePtr = _lookup<
       ffi.NativeFunction<
-          _RoomEventItemItemContentTypeReturn Function(
+          _RoomEventItemEventTypeReturn Function(
     ffi.Int64,
-  )>>("__RoomEventItem_item_content_type");
+  )>>("__RoomEventItem_event_type");
 
-  late final _roomEventItemItemContentType =
-      _roomEventItemItemContentTypePtr.asFunction<
-          _RoomEventItemItemContentTypeReturn Function(
+  late final _roomEventItemEventType = _roomEventItemEventTypePtr.asFunction<
+      _RoomEventItemEventTypeReturn Function(
     int,
   )>();
   late final _roomEventItemMsgtypePtr = _lookup<
@@ -5781,6 +6061,16 @@ class Api {
 
   late final _roomEventItemImageDesc = _roomEventItemImageDescPtr.asFunction<
       _RoomEventItemImageDescReturn Function(
+    int,
+  )>();
+  late final _roomEventItemVideoDescPtr = _lookup<
+      ffi.NativeFunction<
+          _RoomEventItemVideoDescReturn Function(
+    ffi.Int64,
+  )>>("__RoomEventItem_video_desc");
+
+  late final _roomEventItemVideoDesc = _roomEventItemVideoDescPtr.asFunction<
+      _RoomEventItemVideoDescReturn Function(
     int,
   )>();
   late final _roomEventItemFileDescPtr = _lookup<
@@ -5839,6 +6129,27 @@ class Api {
 
   late final _roomEventItemIsEditable = _roomEventItemIsEditablePtr.asFunction<
       int Function(
+    int,
+  )>();
+  late final _roomVirtualItemEventTypePtr = _lookup<
+      ffi.NativeFunction<
+          _RoomVirtualItemEventTypeReturn Function(
+    ffi.Int64,
+  )>>("__RoomVirtualItem_event_type");
+
+  late final _roomVirtualItemEventType =
+      _roomVirtualItemEventTypePtr.asFunction<
+          _RoomVirtualItemEventTypeReturn Function(
+    int,
+  )>();
+  late final _roomVirtualItemDescPtr = _lookup<
+      ffi.NativeFunction<
+          _RoomVirtualItemDescReturn Function(
+    ffi.Int64,
+  )>>("__RoomVirtualItem_desc");
+
+  late final _roomVirtualItemDesc = _roomVirtualItemDescPtr.asFunction<
+      _RoomVirtualItemDescReturn Function(
     int,
   )>();
   late final _roomMessageItemTypePtr = _lookup<
@@ -5951,6 +6262,182 @@ class Api {
       _ImageDescHeightReturn Function(
     int,
   )>();
+  late final _imageDescThumbnailMimetypePtr = _lookup<
+      ffi.NativeFunction<
+          _ImageDescThumbnailMimetypeReturn Function(
+    ffi.Int64,
+  )>>("__ImageDesc_thumbnail_mimetype");
+
+  late final _imageDescThumbnailMimetype =
+      _imageDescThumbnailMimetypePtr.asFunction<
+          _ImageDescThumbnailMimetypeReturn Function(
+    int,
+  )>();
+  late final _imageDescThumbnailSizePtr = _lookup<
+      ffi.NativeFunction<
+          _ImageDescThumbnailSizeReturn Function(
+    ffi.Int64,
+  )>>("__ImageDesc_thumbnail_size");
+
+  late final _imageDescThumbnailSize = _imageDescThumbnailSizePtr.asFunction<
+      _ImageDescThumbnailSizeReturn Function(
+    int,
+  )>();
+  late final _imageDescThumbnailWidthPtr = _lookup<
+      ffi.NativeFunction<
+          _ImageDescThumbnailWidthReturn Function(
+    ffi.Int64,
+  )>>("__ImageDesc_thumbnail_width");
+
+  late final _imageDescThumbnailWidth = _imageDescThumbnailWidthPtr.asFunction<
+      _ImageDescThumbnailWidthReturn Function(
+    int,
+  )>();
+  late final _imageDescThumbnailHeightPtr = _lookup<
+      ffi.NativeFunction<
+          _ImageDescThumbnailHeightReturn Function(
+    ffi.Int64,
+  )>>("__ImageDesc_thumbnail_height");
+
+  late final _imageDescThumbnailHeight =
+      _imageDescThumbnailHeightPtr.asFunction<
+          _ImageDescThumbnailHeightReturn Function(
+    int,
+  )>();
+  late final _imageDescThumbnailSourcePtr = _lookup<
+      ffi.NativeFunction<
+          _ImageDescThumbnailSourceReturn Function(
+    ffi.Int64,
+  )>>("__ImageDesc_thumbnail_source");
+
+  late final _imageDescThumbnailSource =
+      _imageDescThumbnailSourcePtr.asFunction<
+          _ImageDescThumbnailSourceReturn Function(
+    int,
+  )>();
+  late final _videoDescNamePtr = _lookup<
+      ffi.NativeFunction<
+          _VideoDescNameReturn Function(
+    ffi.Int64,
+  )>>("__VideoDesc_name");
+
+  late final _videoDescName = _videoDescNamePtr.asFunction<
+      _VideoDescNameReturn Function(
+    int,
+  )>();
+  late final _videoDescMimetypePtr = _lookup<
+      ffi.NativeFunction<
+          _VideoDescMimetypeReturn Function(
+    ffi.Int64,
+  )>>("__VideoDesc_mimetype");
+
+  late final _videoDescMimetype = _videoDescMimetypePtr.asFunction<
+      _VideoDescMimetypeReturn Function(
+    int,
+  )>();
+  late final _videoDescSizePtr = _lookup<
+      ffi.NativeFunction<
+          _VideoDescSizeReturn Function(
+    ffi.Int64,
+  )>>("__VideoDesc_size");
+
+  late final _videoDescSize = _videoDescSizePtr.asFunction<
+      _VideoDescSizeReturn Function(
+    int,
+  )>();
+  late final _videoDescWidthPtr = _lookup<
+      ffi.NativeFunction<
+          _VideoDescWidthReturn Function(
+    ffi.Int64,
+  )>>("__VideoDesc_width");
+
+  late final _videoDescWidth = _videoDescWidthPtr.asFunction<
+      _VideoDescWidthReturn Function(
+    int,
+  )>();
+  late final _videoDescHeightPtr = _lookup<
+      ffi.NativeFunction<
+          _VideoDescHeightReturn Function(
+    ffi.Int64,
+  )>>("__VideoDesc_height");
+
+  late final _videoDescHeight = _videoDescHeightPtr.asFunction<
+      _VideoDescHeightReturn Function(
+    int,
+  )>();
+  late final _videoDescBlurhashPtr = _lookup<
+      ffi.NativeFunction<
+          _VideoDescBlurhashReturn Function(
+    ffi.Int64,
+  )>>("__VideoDesc_blurhash");
+
+  late final _videoDescBlurhash = _videoDescBlurhashPtr.asFunction<
+      _VideoDescBlurhashReturn Function(
+    int,
+  )>();
+  late final _videoDescDurationPtr = _lookup<
+      ffi.NativeFunction<
+          _VideoDescDurationReturn Function(
+    ffi.Int64,
+  )>>("__VideoDesc_duration");
+
+  late final _videoDescDuration = _videoDescDurationPtr.asFunction<
+      _VideoDescDurationReturn Function(
+    int,
+  )>();
+  late final _videoDescThumbnailMimetypePtr = _lookup<
+      ffi.NativeFunction<
+          _VideoDescThumbnailMimetypeReturn Function(
+    ffi.Int64,
+  )>>("__VideoDesc_thumbnail_mimetype");
+
+  late final _videoDescThumbnailMimetype =
+      _videoDescThumbnailMimetypePtr.asFunction<
+          _VideoDescThumbnailMimetypeReturn Function(
+    int,
+  )>();
+  late final _videoDescThumbnailSizePtr = _lookup<
+      ffi.NativeFunction<
+          _VideoDescThumbnailSizeReturn Function(
+    ffi.Int64,
+  )>>("__VideoDesc_thumbnail_size");
+
+  late final _videoDescThumbnailSize = _videoDescThumbnailSizePtr.asFunction<
+      _VideoDescThumbnailSizeReturn Function(
+    int,
+  )>();
+  late final _videoDescThumbnailWidthPtr = _lookup<
+      ffi.NativeFunction<
+          _VideoDescThumbnailWidthReturn Function(
+    ffi.Int64,
+  )>>("__VideoDesc_thumbnail_width");
+
+  late final _videoDescThumbnailWidth = _videoDescThumbnailWidthPtr.asFunction<
+      _VideoDescThumbnailWidthReturn Function(
+    int,
+  )>();
+  late final _videoDescThumbnailHeightPtr = _lookup<
+      ffi.NativeFunction<
+          _VideoDescThumbnailHeightReturn Function(
+    ffi.Int64,
+  )>>("__VideoDesc_thumbnail_height");
+
+  late final _videoDescThumbnailHeight =
+      _videoDescThumbnailHeightPtr.asFunction<
+          _VideoDescThumbnailHeightReturn Function(
+    int,
+  )>();
+  late final _videoDescThumbnailSourcePtr = _lookup<
+      ffi.NativeFunction<
+          _VideoDescThumbnailSourceReturn Function(
+    ffi.Int64,
+  )>>("__VideoDesc_thumbnail_source");
+
+  late final _videoDescThumbnailSource =
+      _videoDescThumbnailSourcePtr.asFunction<
+          _VideoDescThumbnailSourceReturn Function(
+    int,
+  )>();
   late final _fileDescNamePtr = _lookup<
       ffi.NativeFunction<
           _FileDescNameReturn Function(
@@ -5979,6 +6466,57 @@ class Api {
 
   late final _fileDescSize = _fileDescSizePtr.asFunction<
       _FileDescSizeReturn Function(
+    int,
+  )>();
+  late final _fileDescThumbnailMimetypePtr = _lookup<
+      ffi.NativeFunction<
+          _FileDescThumbnailMimetypeReturn Function(
+    ffi.Int64,
+  )>>("__FileDesc_thumbnail_mimetype");
+
+  late final _fileDescThumbnailMimetype =
+      _fileDescThumbnailMimetypePtr.asFunction<
+          _FileDescThumbnailMimetypeReturn Function(
+    int,
+  )>();
+  late final _fileDescThumbnailSizePtr = _lookup<
+      ffi.NativeFunction<
+          _FileDescThumbnailSizeReturn Function(
+    ffi.Int64,
+  )>>("__FileDesc_thumbnail_size");
+
+  late final _fileDescThumbnailSize = _fileDescThumbnailSizePtr.asFunction<
+      _FileDescThumbnailSizeReturn Function(
+    int,
+  )>();
+  late final _fileDescThumbnailWidthPtr = _lookup<
+      ffi.NativeFunction<
+          _FileDescThumbnailWidthReturn Function(
+    ffi.Int64,
+  )>>("__FileDesc_thumbnail_width");
+
+  late final _fileDescThumbnailWidth = _fileDescThumbnailWidthPtr.asFunction<
+      _FileDescThumbnailWidthReturn Function(
+    int,
+  )>();
+  late final _fileDescThumbnailHeightPtr = _lookup<
+      ffi.NativeFunction<
+          _FileDescThumbnailHeightReturn Function(
+    ffi.Int64,
+  )>>("__FileDesc_thumbnail_height");
+
+  late final _fileDescThumbnailHeight = _fileDescThumbnailHeightPtr.asFunction<
+      _FileDescThumbnailHeightReturn Function(
+    int,
+  )>();
+  late final _fileDescThumbnailSourcePtr = _lookup<
+      ffi.NativeFunction<
+          _FileDescThumbnailSourceReturn Function(
+    ffi.Int64,
+  )>>("__FileDesc_thumbnail_source");
+
+  late final _fileDescThumbnailSource = _fileDescThumbnailSourcePtr.asFunction<
+      _FileDescThumbnailSourceReturn Function(
     int,
   )>();
   late final _reactionDescCountPtr = _lookup<
@@ -8034,6 +8572,16 @@ class Api {
       int Function(
     int,
   )>();
+  late final _taskListTaskBuilderPtr = _lookup<
+      ffi.NativeFunction<
+          _TaskListTaskBuilderReturn Function(
+    ffi.Int64,
+  )>>("__TaskList_task_builder");
+
+  late final _taskListTaskBuilder = _taskListTaskBuilderPtr.asFunction<
+      _TaskListTaskBuilderReturn Function(
+    int,
+  )>();
   late final _taskListUpdateBuilderPtr = _lookup<
       ffi.NativeFunction<
           _TaskListUpdateBuilderReturn Function(
@@ -8462,6 +9010,16 @@ class Api {
       int Function(
     int,
   )>();
+  late final _groupGetRoomIdPtr = _lookup<
+      ffi.NativeFunction<
+          _GroupGetRoomIdReturn Function(
+    ffi.Int64,
+  )>>("__Group_get_room_id");
+
+  late final _groupGetRoomId = _groupGetRoomIdPtr.asFunction<
+      _GroupGetRoomIdReturn Function(
+    int,
+  )>();
   late final _groupGetMemberPtr = _lookup<
       ffi.NativeFunction<
           ffi.Int64 Function(
@@ -8614,6 +9172,57 @@ class Api {
 
   late final _syncStateCancel = _syncStateCancelPtr.asFunction<
       void Function(
+    int,
+  )>();
+  late final _createGroupSettingsAliasPtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Void Function(
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Uint64,
+    ffi.Uint64,
+  )>>("__CreateGroupSettings_alias");
+
+  late final _createGroupSettingsAlias =
+      _createGroupSettingsAliasPtr.asFunction<
+          void Function(
+    int,
+    int,
+    int,
+    int,
+  )>();
+  late final _createGroupSettingsVisibilityPtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Void Function(
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Uint64,
+    ffi.Uint64,
+  )>>("__CreateGroupSettings_visibility");
+
+  late final _createGroupSettingsVisibility =
+      _createGroupSettingsVisibilityPtr.asFunction<
+          void Function(
+    int,
+    int,
+    int,
+    int,
+  )>();
+  late final _createGroupSettingsAddInviteePtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Void Function(
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Uint64,
+    ffi.Uint64,
+  )>>("__CreateGroupSettings_add_invitee");
+
+  late final _createGroupSettingsAddInvitee =
+      _createGroupSettingsAddInviteePtr.asFunction<
+          void Function(
+    int,
+    int,
+    int,
     int,
   )>();
   late final _clientStartSyncPtr = _lookup<
@@ -8905,6 +9514,95 @@ class Api {
 
   late final _clientTaskLists = _clientTaskListsPtr.asFunction<
       int Function(
+    int,
+  )>();
+  late final _clientCreateEffektioGroupPtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Int64 Function(
+    ffi.Int64,
+    ffi.Int64,
+  )>>("__Client_create_effektio_group");
+
+  late final _clientCreateEffektioGroup =
+      _clientCreateEffektioGroupPtr.asFunction<
+          int Function(
+    int,
+    int,
+  )>();
+  late final _clientSubscribePtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Int64 Function(
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Uint64,
+    ffi.Uint64,
+  )>>("__Client_subscribe");
+
+  late final _clientSubscribe = _clientSubscribePtr.asFunction<
+      int Function(
+    int,
+    int,
+    int,
+    int,
+  )>();
+  late final _clientWaitForCommentPtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Int64 Function(
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Uint64,
+    ffi.Uint64,
+    ffi.Uint8,
+    ffi.Int64,
+  )>>("__Client_wait_for_comment");
+
+  late final _clientWaitForComment = _clientWaitForCommentPtr.asFunction<
+      int Function(
+    int,
+    int,
+    int,
+    int,
+    int,
+    int,
+  )>();
+  late final _clientWaitForTaskListPtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Int64 Function(
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Uint64,
+    ffi.Uint64,
+    ffi.Uint8,
+    ffi.Int64,
+  )>>("__Client_wait_for_task_list");
+
+  late final _clientWaitForTaskList = _clientWaitForTaskListPtr.asFunction<
+      int Function(
+    int,
+    int,
+    int,
+    int,
+    int,
+    int,
+  )>();
+  late final _clientWaitForTaskPtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Int64 Function(
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Uint64,
+    ffi.Uint64,
+    ffi.Uint8,
+    ffi.Int64,
+  )>>("__Client_wait_for_task");
+
+  late final _clientWaitForTask = _clientWaitForTaskPtr.asFunction<
+      int Function(
+    int,
+    int,
+    int,
+    int,
+    int,
     int,
   )>();
   late final _userProfileUserIdPtr = _lookup<
@@ -10480,6 +11178,66 @@ class Api {
     int,
     int,
   )>();
+  late final _clientCreateEffektioGroupFuturePollPtr = _lookup<
+      ffi.NativeFunction<
+          _ClientCreateEffektioGroupFuturePollReturn Function(
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Int64,
+  )>>("__Client_create_effektio_group_future_poll");
+
+  late final _clientCreateEffektioGroupFuturePoll =
+      _clientCreateEffektioGroupFuturePollPtr.asFunction<
+          _ClientCreateEffektioGroupFuturePollReturn Function(
+    int,
+    int,
+    int,
+  )>();
+  late final _clientWaitForCommentFuturePollPtr = _lookup<
+      ffi.NativeFunction<
+          _ClientWaitForCommentFuturePollReturn Function(
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Int64,
+  )>>("__Client_wait_for_comment_future_poll");
+
+  late final _clientWaitForCommentFuturePoll =
+      _clientWaitForCommentFuturePollPtr.asFunction<
+          _ClientWaitForCommentFuturePollReturn Function(
+    int,
+    int,
+    int,
+  )>();
+  late final _clientWaitForTaskListFuturePollPtr = _lookup<
+      ffi.NativeFunction<
+          _ClientWaitForTaskListFuturePollReturn Function(
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Int64,
+  )>>("__Client_wait_for_task_list_future_poll");
+
+  late final _clientWaitForTaskListFuturePoll =
+      _clientWaitForTaskListFuturePollPtr.asFunction<
+          _ClientWaitForTaskListFuturePollReturn Function(
+    int,
+    int,
+    int,
+  )>();
+  late final _clientWaitForTaskFuturePollPtr = _lookup<
+      ffi.NativeFunction<
+          _ClientWaitForTaskFuturePollReturn Function(
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Int64,
+  )>>("__Client_wait_for_task_future_poll");
+
+  late final _clientWaitForTaskFuturePoll =
+      _clientWaitForTaskFuturePollPtr.asFunction<
+          _ClientWaitForTaskFuturePollReturn Function(
+    int,
+    int,
+    int,
+  )>();
   late final _userProfileGetAvatarFuturePollPtr = _lookup<
       ffi.NativeFunction<
           _UserProfileGetAvatarFuturePollReturn Function(
@@ -11040,6 +11798,23 @@ class Api {
   late final _clientIncomingMessageRxStreamPoll =
       _clientIncomingMessageRxStreamPollPtr.asFunction<
           _ClientIncomingMessageRxStreamPollReturn Function(
+    int,
+    int,
+    int,
+    int,
+  )>();
+  late final _clientSubscribeStreamPollPtr = _lookup<
+      ffi.NativeFunction<
+          _ClientSubscribeStreamPollReturn Function(
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Int64,
+  )>>("__Client_subscribe_stream_poll");
+
+  late final _clientSubscribeStreamPoll =
+      _clientSubscribeStreamPollPtr.asFunction<
+          _ClientSubscribeStreamPollReturn Function(
     int,
     int,
     int,
@@ -11925,18 +12700,31 @@ class Api {
           .asFunction<void Function(int, int, int)>();
 }
 
-/// Representing a color
-class Color {
+/// Representing a time frame
+class EfkDuration {
   final Api _api;
   final _Box _box;
 
-  Color._(this._api, this._box);
+  EfkDuration._(this._api, this._box);
+
+  /// Manually drops the object and unregisters the FinalizableHandle.
+  void drop() {
+    _box.drop();
+  }
+}
+
+/// Representing a color
+class EfkColor {
+  final Api _api;
+  final _Box _box;
+
+  EfkColor._(this._api, this._box);
 
   /// as rgba in u8
   List<dynamic> rgbaU8() {
     var tmp0 = 0;
     tmp0 = _box.borrow();
-    final tmp1 = _api._colorRgbaU8(
+    final tmp1 = _api._efkColorRgbaU8(
       tmp0,
     );
     final tmp4 = tmp1.arg0;
@@ -12114,7 +12902,7 @@ class News {
   }
 
   /// if given, the specific foreground color
-  Color? fgColor() {
+  EfkColor? fgColor() {
     var tmp0 = 0;
     tmp0 = _box.borrow();
     final tmp1 = _api._newsFgColor(
@@ -12126,14 +12914,14 @@ class News {
       return null;
     }
     final ffi.Pointer<ffi.Void> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
-    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_Color");
+    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_EfkColor");
     tmp4_1._finalizer = _api._registerFinalizer(tmp4_1);
-    final tmp2 = Color._(_api, tmp4_1);
+    final tmp2 = EfkColor._(_api, tmp4_1);
     return tmp2;
   }
 
   /// if given, the specific background color
-  Color? bgColor() {
+  EfkColor? bgColor() {
     var tmp0 = 0;
     tmp0 = _box.borrow();
     final tmp1 = _api._newsBgColor(
@@ -12145,9 +12933,9 @@ class News {
       return null;
     }
     final ffi.Pointer<ffi.Void> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
-    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_Color");
+    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_EfkColor");
     tmp4_1._finalizer = _api._registerFinalizer(tmp4_1);
-    final tmp2 = Color._(_api, tmp4_1);
+    final tmp2 = EfkColor._(_api, tmp4_1);
     return tmp2;
   }
 
@@ -12228,7 +13016,7 @@ class Tag {
   }
 
   /// if given, the specific color for this tag
-  Color? color() {
+  EfkColor? color() {
     var tmp0 = 0;
     tmp0 = _box.borrow();
     final tmp1 = _api._tagColor(
@@ -12240,9 +13028,9 @@ class Tag {
       return null;
     }
     final ffi.Pointer<ffi.Void> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
-    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_Color");
+    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_EfkColor");
     tmp4_1._finalizer = _api._registerFinalizer(tmp4_1);
-    final tmp2 = Color._(_api, tmp4_1);
+    final tmp2 = EfkColor._(_api, tmp4_1);
     return tmp2;
   }
 
@@ -12350,6 +13138,18 @@ class Faq {
     final tmp2 = tmp3;
     return tmp2;
   }
+
+  /// Manually drops the object and unregisters the FinalizableHandle.
+  void drop() {
+    _box.drop();
+  }
+}
+
+class MediaSource {
+  final Api _api;
+  final _Box _box;
+
+  MediaSource._(this._api, this._box);
 
   /// Manually drops the object and unregisters the FinalizableHandle.
   void drop() {
@@ -12540,11 +13340,11 @@ class RoomEventItem {
     return tmp2;
   }
 
-  /// one of Message/RedactedMessage/UnableToDecrypt/FailedToParseMessageLike/FailedToParseState
-  String itemContentType() {
+  /// one of Message/Redaction/UnableToDecrypt/FailedToParseMessageLike/FailedToParseState
+  String eventType() {
     var tmp0 = 0;
     tmp0 = _box.borrow();
-    final tmp1 = _api._roomEventItemItemContentType(
+    final tmp1 = _api._roomEventItemEventType(
       tmp0,
     );
     final tmp3 = tmp1.arg0;
@@ -12619,6 +13419,25 @@ class RoomEventItem {
     final tmp4_1 = _Box(_api, tmp4_0, "drop_box_ImageDesc");
     tmp4_1._finalizer = _api._registerFinalizer(tmp4_1);
     final tmp2 = ImageDesc._(_api, tmp4_1);
+    return tmp2;
+  }
+
+  /// contains source data, name, mimetype, size, width and height
+  VideoDesc? videoDesc() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._roomEventItemVideoDesc(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final ffi.Pointer<ffi.Void> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_VideoDesc");
+    tmp4_1._finalizer = _api._registerFinalizer(tmp4_1);
+    final tmp2 = VideoDesc._(_api, tmp4_1);
     return tmp2;
   }
 
@@ -12739,6 +13558,50 @@ class RoomVirtualItem {
   final _Box _box;
 
   RoomVirtualItem._(this._api, this._box);
+
+  /// one of DayDivider/LoadingIndicator/ReadMarker/TimelineStart
+  String eventType() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._roomVirtualItemEventType(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    final tmp5 = tmp1.arg2;
+    final ffi.Pointer<ffi.Uint8> tmp3_0 = ffi.Pointer.fromAddress(tmp3);
+    final tmp2 = utf8.decode(tmp3_0.asTypedList(tmp4));
+    if (tmp5 > 0) {
+      final ffi.Pointer<ffi.Void> tmp3_0;
+      tmp3_0 = ffi.Pointer.fromAddress(tmp3);
+      _api.__deallocate(tmp3_0, tmp5 * 1, 1);
+    }
+    return tmp2;
+  }
+
+  /// contains description text
+  String? desc() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._roomVirtualItemDesc(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    final tmp5 = tmp1.arg2;
+    final tmp6 = tmp1.arg3;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final ffi.Pointer<ffi.Uint8> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+    final tmp2 = utf8.decode(tmp4_0.asTypedList(tmp5));
+    if (tmp6 > 0) {
+      final ffi.Pointer<ffi.Void> tmp4_0;
+      tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+      _api.__deallocate(tmp4_0, tmp6 * 1, 1);
+    }
+    return tmp2;
+  }
 
   /// Manually drops the object and unregisters the FinalizableHandle.
   void drop() {
@@ -12991,6 +13854,332 @@ class ImageDesc {
     return tmp2;
   }
 
+  /// thumbnail mimetype
+  String? thumbnailMimetype() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._imageDescThumbnailMimetype(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    final tmp5 = tmp1.arg2;
+    final tmp6 = tmp1.arg3;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final ffi.Pointer<ffi.Uint8> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+    final tmp2 = utf8.decode(tmp4_0.asTypedList(tmp5));
+    if (tmp6 > 0) {
+      final ffi.Pointer<ffi.Void> tmp4_0;
+      tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+      _api.__deallocate(tmp4_0, tmp6 * 1, 1);
+    }
+    return tmp2;
+  }
+
+  /// thumbnail file size
+  int? thumbnailSize() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._imageDescThumbnailSize(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// thumbnail image width
+  int? thumbnailWidth() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._imageDescThumbnailWidth(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// thumbnail image height
+  int? thumbnailHeight() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._imageDescThumbnailHeight(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// thumbnail source
+  MediaSource? thumbnailSource() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._imageDescThumbnailSource(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final ffi.Pointer<ffi.Void> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_MediaSource");
+    tmp4_1._finalizer = _api._registerFinalizer(tmp4_1);
+    final tmp2 = MediaSource._(_api, tmp4_1);
+    return tmp2;
+  }
+
+  /// Manually drops the object and unregisters the FinalizableHandle.
+  void drop() {
+    _box.drop();
+  }
+}
+
+class VideoDesc {
+  final Api _api;
+  final _Box _box;
+
+  VideoDesc._(this._api, this._box);
+
+  /// file name
+  String name() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._videoDescName(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    final tmp5 = tmp1.arg2;
+    final ffi.Pointer<ffi.Uint8> tmp3_0 = ffi.Pointer.fromAddress(tmp3);
+    final tmp2 = utf8.decode(tmp3_0.asTypedList(tmp4));
+    if (tmp5 > 0) {
+      final ffi.Pointer<ffi.Void> tmp3_0;
+      tmp3_0 = ffi.Pointer.fromAddress(tmp3);
+      _api.__deallocate(tmp3_0, tmp5 * 1, 1);
+    }
+    return tmp2;
+  }
+
+  /// MIME
+  String? mimetype() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._videoDescMimetype(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    final tmp5 = tmp1.arg2;
+    final tmp6 = tmp1.arg3;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final ffi.Pointer<ffi.Uint8> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+    final tmp2 = utf8.decode(tmp4_0.asTypedList(tmp5));
+    if (tmp6 > 0) {
+      final ffi.Pointer<ffi.Void> tmp4_0;
+      tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+      _api.__deallocate(tmp4_0, tmp6 * 1, 1);
+    }
+    return tmp2;
+  }
+
+  /// file size in bytes
+  int? size() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._videoDescSize(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// image width
+  int? width() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._videoDescWidth(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// image height
+  int? height() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._videoDescHeight(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// blurhash
+  String? blurhash() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._videoDescBlurhash(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    final tmp5 = tmp1.arg2;
+    final tmp6 = tmp1.arg3;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final ffi.Pointer<ffi.Uint8> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+    final tmp2 = utf8.decode(tmp4_0.asTypedList(tmp5));
+    if (tmp6 > 0) {
+      final ffi.Pointer<ffi.Void> tmp4_0;
+      tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+      _api.__deallocate(tmp4_0, tmp6 * 1, 1);
+    }
+    return tmp2;
+  }
+
+  /// duration in seconds
+  int? duration() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._videoDescDuration(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// thumbnail mimetype
+  String? thumbnailMimetype() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._videoDescThumbnailMimetype(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    final tmp5 = tmp1.arg2;
+    final tmp6 = tmp1.arg3;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final ffi.Pointer<ffi.Uint8> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+    final tmp2 = utf8.decode(tmp4_0.asTypedList(tmp5));
+    if (tmp6 > 0) {
+      final ffi.Pointer<ffi.Void> tmp4_0;
+      tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+      _api.__deallocate(tmp4_0, tmp6 * 1, 1);
+    }
+    return tmp2;
+  }
+
+  /// thumbnail file size
+  int? thumbnailSize() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._videoDescThumbnailSize(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// thumbnail image width
+  int? thumbnailWidth() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._videoDescThumbnailWidth(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// thumbnail image height
+  int? thumbnailHeight() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._videoDescThumbnailHeight(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// thumbnail source
+  MediaSource? thumbnailSource() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._videoDescThumbnailSource(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final ffi.Pointer<ffi.Void> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_MediaSource");
+    tmp4_1._finalizer = _api._registerFinalizer(tmp4_1);
+    final tmp2 = MediaSource._(_api, tmp4_1);
+    return tmp2;
+  }
+
   /// Manually drops the object and unregisters the FinalizableHandle.
   void drop() {
     _box.drop();
@@ -13060,6 +14249,97 @@ class FileDesc {
       return null;
     }
     final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// thumbnail mimetype
+  String? thumbnailMimetype() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._fileDescThumbnailMimetype(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    final tmp5 = tmp1.arg2;
+    final tmp6 = tmp1.arg3;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final ffi.Pointer<ffi.Uint8> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+    final tmp2 = utf8.decode(tmp4_0.asTypedList(tmp5));
+    if (tmp6 > 0) {
+      final ffi.Pointer<ffi.Void> tmp4_0;
+      tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+      _api.__deallocate(tmp4_0, tmp6 * 1, 1);
+    }
+    return tmp2;
+  }
+
+  /// thumbnail file size
+  int? thumbnailSize() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._fileDescThumbnailSize(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// thumbnail image width
+  int? thumbnailWidth() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._fileDescThumbnailWidth(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// thumbnail image height
+  int? thumbnailHeight() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._fileDescThumbnailHeight(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final tmp2 = tmp4;
+    return tmp2;
+  }
+
+  /// thumbnail source
+  MediaSource? thumbnailSource() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._fileDescThumbnailSource(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    if (tmp3 == 0) {
+      return null;
+    }
+    final ffi.Pointer<ffi.Void> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_MediaSource");
+    tmp4_1._finalizer = _api._registerFinalizer(tmp4_1);
+    final tmp2 = MediaSource._(_api, tmp4_1);
     return tmp2;
   }
 
@@ -15262,7 +16542,7 @@ class Task {
   }
 
   /// Has this been colored in?
-  Color? color() {
+  EfkColor? color() {
     var tmp0 = 0;
     tmp0 = _box.borrow();
     final tmp1 = _api._taskColor(
@@ -15274,9 +16554,9 @@ class Task {
       return null;
     }
     final ffi.Pointer<ffi.Void> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
-    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_Color");
+    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_EfkColor");
     tmp4_1._finalizer = _api._registerFinalizer(tmp4_1);
-    final tmp2 = Color._(_api, tmp4_1);
+    final tmp2 = EfkColor._(_api, tmp4_1);
     return tmp2;
   }
 
@@ -15532,7 +16812,7 @@ class TaskUpdateBuilder {
 
   /// set the color for this task list
   void color(
-    Color color,
+    EfkColor color,
   ) {
     final tmp1 = color;
     var tmp0 = 0;
@@ -16179,7 +17459,7 @@ class TaskDraft {
 
   /// set the color for this task list
   void color(
-    Color color,
+    EfkColor color,
   ) {
     final tmp1 = color;
     var tmp0 = 0;
@@ -16752,7 +18032,7 @@ class TaskList {
   }
 
   /// Has this been colored in?
-  Color? color() {
+  EfkColor? color() {
     var tmp0 = 0;
     tmp0 = _box.borrow();
     final tmp1 = _api._taskListColor(
@@ -16764,9 +18044,9 @@ class TaskList {
       return null;
     }
     final ffi.Pointer<ffi.Void> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
-    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_Color");
+    final tmp4_1 = _Box(_api, tmp4_0, "drop_box_EfkColor");
     tmp4_1._finalizer = _api._registerFinalizer(tmp4_1);
-    final tmp2 = Color._(_api, tmp4_1);
+    final tmp2 = EfkColor._(_api, tmp4_1);
     return tmp2;
   }
 
@@ -16838,6 +18118,35 @@ class TaskList {
     final tmp3_1 = _Box(_api, tmp3_0, "__TaskList_tasks_future_drop");
     tmp3_1._finalizer = _api._registerFinalizer(tmp3_1);
     final tmp2 = _nativeFuture(tmp3_1, _api.__taskListTasksFuturePoll);
+    return tmp2;
+  }
+
+  /// make a builder for creating the task draft
+  TaskDraft taskBuilder() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._taskListTaskBuilder(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    final tmp5 = tmp1.arg2;
+    final tmp6 = tmp1.arg3;
+    final tmp7 = tmp1.arg4;
+    if (tmp3 == 0) {
+      final ffi.Pointer<ffi.Uint8> tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+      final tmp3_0 = utf8.decode(tmp4_0.asTypedList(tmp5));
+      if (tmp5 > 0) {
+        final ffi.Pointer<ffi.Void> tmp4_0;
+        tmp4_0 = ffi.Pointer.fromAddress(tmp4);
+        _api.__deallocate(tmp4_0, tmp6, 1);
+      }
+      throw tmp3_0;
+    }
+    final ffi.Pointer<ffi.Void> tmp7_0 = ffi.Pointer.fromAddress(tmp7);
+    final tmp7_1 = _Box(_api, tmp7_0, "drop_box_TaskDraft");
+    tmp7_1._finalizer = _api._registerFinalizer(tmp7_1);
+    final tmp2 = TaskDraft._(_api, tmp7_1);
     return tmp2;
   }
 
@@ -16990,7 +18299,7 @@ class TaskListDraft {
 
   /// set the color for this task list
   void color(
-    Color color,
+    EfkColor color,
   ) {
     final tmp1 = color;
     var tmp0 = 0;
@@ -17203,7 +18512,7 @@ class TaskListUpdateBuilder {
 
   /// set the color for this task list
   void color(
-    Color color,
+    EfkColor color,
   ) {
     final tmp1 = color;
     var tmp0 = 0;
@@ -17393,6 +18702,26 @@ class Group {
     final tmp3_1 = _Box(_api, tmp3_0, "__Group_active_members_future_drop");
     tmp3_1._finalizer = _api._registerFinalizer(tmp3_1);
     final tmp2 = _nativeFuture(tmp3_1, _api.__groupActiveMembersFuturePoll);
+    return tmp2;
+  }
+
+  /// the room id
+  String getRoomId() {
+    var tmp0 = 0;
+    tmp0 = _box.borrow();
+    final tmp1 = _api._groupGetRoomId(
+      tmp0,
+    );
+    final tmp3 = tmp1.arg0;
+    final tmp4 = tmp1.arg1;
+    final tmp5 = tmp1.arg2;
+    final ffi.Pointer<ffi.Uint8> tmp3_0 = ffi.Pointer.fromAddress(tmp3);
+    final tmp2 = utf8.decode(tmp3_0.asTypedList(tmp4));
+    if (tmp5 > 0) {
+      final ffi.Pointer<ffi.Void> tmp3_0;
+      tmp3_0 = ffi.Pointer.fromAddress(tmp3);
+      _api.__deallocate(tmp3_0, tmp5 * 1, 1);
+    }
     return tmp2;
   }
 
@@ -17709,6 +19038,96 @@ class SyncState {
     tmp0 = _box.borrow();
     _api._syncStateCancel(
       tmp0,
+    );
+    return;
+  }
+
+  /// Manually drops the object and unregisters the FinalizableHandle.
+  void drop() {
+    _box.drop();
+  }
+}
+
+class CreateGroupSettings {
+  final Api _api;
+  final _Box _box;
+
+  CreateGroupSettings._(this._api, this._box);
+
+  /// set the alias of group
+  void alias(
+    String value,
+  ) {
+    final tmp1 = value;
+    var tmp0 = 0;
+    var tmp2 = 0;
+    var tmp3 = 0;
+    var tmp4 = 0;
+    tmp0 = _box.borrow();
+    final tmp1_0 = utf8.encode(tmp1);
+    tmp3 = tmp1_0.length;
+    final ffi.Pointer<ffi.Uint8> tmp2_0 = _api.__allocate(tmp3 * 1, 1);
+    final Uint8List tmp2_1 = tmp2_0.asTypedList(tmp3);
+    tmp2_1.setAll(0, tmp1_0);
+    tmp2 = tmp2_0.address;
+    tmp4 = tmp3;
+    _api._createGroupSettingsAlias(
+      tmp0,
+      tmp2,
+      tmp3,
+      tmp4,
+    );
+    return;
+  }
+
+  /// set the group's visibility to either Public or Private
+  void visibility(
+    String value,
+  ) {
+    final tmp1 = value;
+    var tmp0 = 0;
+    var tmp2 = 0;
+    var tmp3 = 0;
+    var tmp4 = 0;
+    tmp0 = _box.borrow();
+    final tmp1_0 = utf8.encode(tmp1);
+    tmp3 = tmp1_0.length;
+    final ffi.Pointer<ffi.Uint8> tmp2_0 = _api.__allocate(tmp3 * 1, 1);
+    final Uint8List tmp2_1 = tmp2_0.asTypedList(tmp3);
+    tmp2_1.setAll(0, tmp1_0);
+    tmp2 = tmp2_0.address;
+    tmp4 = tmp3;
+    _api._createGroupSettingsVisibility(
+      tmp0,
+      tmp2,
+      tmp3,
+      tmp4,
+    );
+    return;
+  }
+
+  /// add the id of user that will be invited to this group
+  void addInvitee(
+    String value,
+  ) {
+    final tmp1 = value;
+    var tmp0 = 0;
+    var tmp2 = 0;
+    var tmp3 = 0;
+    var tmp4 = 0;
+    tmp0 = _box.borrow();
+    final tmp1_0 = utf8.encode(tmp1);
+    tmp3 = tmp1_0.length;
+    final ffi.Pointer<ffi.Uint8> tmp2_0 = _api.__allocate(tmp3 * 1, 1);
+    final Uint8List tmp2_1 = tmp2_0.asTypedList(tmp3);
+    tmp2_1.setAll(0, tmp1_0);
+    tmp2 = tmp2_0.address;
+    tmp4 = tmp3;
+    _api._createGroupSettingsAddInvitee(
+      tmp0,
+      tmp2,
+      tmp3,
+      tmp4,
     );
     return;
   }
@@ -18247,6 +19666,195 @@ class Client {
     tmp3_1._finalizer = _api._registerFinalizer(tmp3_1);
     final tmp2 = _nativeFuture(tmp3_1, _api.__clientTaskListsFuturePoll);
     return tmp2;
+  }
+
+  /// create default group
+  Future<RoomId> createEffektioGroup(
+    CreateGroupSettings settings,
+  ) {
+    final tmp1 = settings;
+    var tmp0 = 0;
+    var tmp2 = 0;
+    tmp0 = _box.borrow();
+    tmp2 = tmp1._box.move();
+    final tmp3 = _api._clientCreateEffektioGroup(
+      tmp0,
+      tmp2,
+    );
+    final tmp5 = tmp3;
+    final ffi.Pointer<ffi.Void> tmp5_0 = ffi.Pointer.fromAddress(tmp5);
+    final tmp5_1 =
+        _Box(_api, tmp5_0, "__Client_create_effektio_group_future_drop");
+    tmp5_1._finalizer = _api._registerFinalizer(tmp5_1);
+    final tmp4 =
+        _nativeFuture(tmp5_1, _api.__clientCreateEffektioGroupFuturePoll);
+    return tmp4;
+  }
+
+  /// listen to updates to any model key
+  Stream<bool> subscribe(
+    String key,
+  ) {
+    final tmp1 = key;
+    var tmp0 = 0;
+    var tmp2 = 0;
+    var tmp3 = 0;
+    var tmp4 = 0;
+    tmp0 = _box.borrow();
+    final tmp1_0 = utf8.encode(tmp1);
+    tmp3 = tmp1_0.length;
+    final ffi.Pointer<ffi.Uint8> tmp2_0 = _api.__allocate(tmp3 * 1, 1);
+    final Uint8List tmp2_1 = tmp2_0.asTypedList(tmp3);
+    tmp2_1.setAll(0, tmp1_0);
+    tmp2 = tmp2_0.address;
+    tmp4 = tmp3;
+    final tmp5 = _api._clientSubscribe(
+      tmp0,
+      tmp2,
+      tmp3,
+      tmp4,
+    );
+    final tmp7 = tmp5;
+    final ffi.Pointer<ffi.Void> tmp7_0 = ffi.Pointer.fromAddress(tmp7);
+    final tmp7_1 = _Box(_api, tmp7_0, "__Client_subscribe_stream_drop");
+    tmp7_1._finalizer = _api._registerFinalizer(tmp7_1);
+    final tmp6 = _nativeStream(tmp7_1, _api.__clientSubscribeStreamPoll);
+    return tmp6;
+  }
+
+  /// Fetch the Comment or use its event_id to wait for it to come down the wire
+  Future<Comment> waitForComment(
+    String key,
+    EfkDuration? timeout,
+  ) {
+    final tmp1 = key;
+    final tmp5 = timeout;
+    var tmp0 = 0;
+    var tmp2 = 0;
+    var tmp3 = 0;
+    var tmp4 = 0;
+    var tmp6 = 0;
+    var tmp8 = 0;
+    tmp0 = _box.borrow();
+    final tmp1_0 = utf8.encode(tmp1);
+    tmp3 = tmp1_0.length;
+    final ffi.Pointer<ffi.Uint8> tmp2_0 = _api.__allocate(tmp3 * 1, 1);
+    final Uint8List tmp2_1 = tmp2_0.asTypedList(tmp3);
+    tmp2_1.setAll(0, tmp1_0);
+    tmp2 = tmp2_0.address;
+    tmp4 = tmp3;
+    if (tmp5 == null) {
+      tmp6 = 0;
+    } else {
+      tmp6 = 1;
+      final tmp7 = tmp5;
+      tmp8 = tmp7._box.move();
+    }
+    final tmp9 = _api._clientWaitForComment(
+      tmp0,
+      tmp2,
+      tmp3,
+      tmp4,
+      tmp6,
+      tmp8,
+    );
+    final tmp11 = tmp9;
+    final ffi.Pointer<ffi.Void> tmp11_0 = ffi.Pointer.fromAddress(tmp11);
+    final tmp11_1 =
+        _Box(_api, tmp11_0, "__Client_wait_for_comment_future_drop");
+    tmp11_1._finalizer = _api._registerFinalizer(tmp11_1);
+    final tmp10 = _nativeFuture(tmp11_1, _api.__clientWaitForCommentFuturePoll);
+    return tmp10;
+  }
+
+  /// Fetch the Tasklist or use its event_id to wait for it to come down the wire
+  Future<TaskList> waitForTaskList(
+    String key,
+    EfkDuration? timeout,
+  ) {
+    final tmp1 = key;
+    final tmp5 = timeout;
+    var tmp0 = 0;
+    var tmp2 = 0;
+    var tmp3 = 0;
+    var tmp4 = 0;
+    var tmp6 = 0;
+    var tmp8 = 0;
+    tmp0 = _box.borrow();
+    final tmp1_0 = utf8.encode(tmp1);
+    tmp3 = tmp1_0.length;
+    final ffi.Pointer<ffi.Uint8> tmp2_0 = _api.__allocate(tmp3 * 1, 1);
+    final Uint8List tmp2_1 = tmp2_0.asTypedList(tmp3);
+    tmp2_1.setAll(0, tmp1_0);
+    tmp2 = tmp2_0.address;
+    tmp4 = tmp3;
+    if (tmp5 == null) {
+      tmp6 = 0;
+    } else {
+      tmp6 = 1;
+      final tmp7 = tmp5;
+      tmp8 = tmp7._box.move();
+    }
+    final tmp9 = _api._clientWaitForTaskList(
+      tmp0,
+      tmp2,
+      tmp3,
+      tmp4,
+      tmp6,
+      tmp8,
+    );
+    final tmp11 = tmp9;
+    final ffi.Pointer<ffi.Void> tmp11_0 = ffi.Pointer.fromAddress(tmp11);
+    final tmp11_1 =
+        _Box(_api, tmp11_0, "__Client_wait_for_task_list_future_drop");
+    tmp11_1._finalizer = _api._registerFinalizer(tmp11_1);
+    final tmp10 =
+        _nativeFuture(tmp11_1, _api.__clientWaitForTaskListFuturePoll);
+    return tmp10;
+  }
+
+  /// Fetch the Task or use its event_id to wait for it to come down the wire
+  Future<Task> waitForTask(
+    String key,
+    EfkDuration? timeout,
+  ) {
+    final tmp1 = key;
+    final tmp5 = timeout;
+    var tmp0 = 0;
+    var tmp2 = 0;
+    var tmp3 = 0;
+    var tmp4 = 0;
+    var tmp6 = 0;
+    var tmp8 = 0;
+    tmp0 = _box.borrow();
+    final tmp1_0 = utf8.encode(tmp1);
+    tmp3 = tmp1_0.length;
+    final ffi.Pointer<ffi.Uint8> tmp2_0 = _api.__allocate(tmp3 * 1, 1);
+    final Uint8List tmp2_1 = tmp2_0.asTypedList(tmp3);
+    tmp2_1.setAll(0, tmp1_0);
+    tmp2 = tmp2_0.address;
+    tmp4 = tmp3;
+    if (tmp5 == null) {
+      tmp6 = 0;
+    } else {
+      tmp6 = 1;
+      final tmp7 = tmp5;
+      tmp8 = tmp7._box.move();
+    }
+    final tmp9 = _api._clientWaitForTask(
+      tmp0,
+      tmp2,
+      tmp3,
+      tmp4,
+      tmp6,
+      tmp8,
+    );
+    final tmp11 = tmp9;
+    final ffi.Pointer<ffi.Void> tmp11_0 = ffi.Pointer.fromAddress(tmp11);
+    final tmp11_1 = _Box(_api, tmp11_0, "__Client_wait_for_task_future_drop");
+    tmp11_1._finalizer = _api._registerFinalizer(tmp11_1);
+    final tmp10 = _nativeFuture(tmp11_1, _api.__clientWaitForTaskFuturePoll);
+    return tmp10;
   }
 
   /// Manually drops the object and unregisters the FinalizableHandle.
@@ -19427,7 +21035,7 @@ class _InitLoggingReturn extends ffi.Struct {
   external int arg3;
 }
 
-class _ColorRgbaU8Return extends ffi.Struct {
+class _EfkColorRgbaU8Return extends ffi.Struct {
   @ffi.Uint8()
   external int arg0;
   @ffi.Uint8()
@@ -19598,7 +21206,7 @@ class _RoomEventItemSenderReturn extends ffi.Struct {
   external int arg2;
 }
 
-class _RoomEventItemItemContentTypeReturn extends ffi.Struct {
+class _RoomEventItemEventTypeReturn extends ffi.Struct {
   @ffi.Int64()
   external int arg0;
   @ffi.Uint64()
@@ -19632,6 +21240,13 @@ class _RoomEventItemImageDescReturn extends ffi.Struct {
   external int arg1;
 }
 
+class _RoomEventItemVideoDescReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Int64()
+  external int arg1;
+}
+
 class _RoomEventItemFileDescReturn extends ffi.Struct {
   @ffi.Uint8()
   external int arg0;
@@ -19655,6 +21270,26 @@ class _RoomEventItemReactionDescReturn extends ffi.Struct {
   external int arg0;
   @ffi.Int64()
   external int arg1;
+}
+
+class _RoomVirtualItemEventTypeReturn extends ffi.Struct {
+  @ffi.Int64()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+  @ffi.Uint64()
+  external int arg2;
+}
+
+class _RoomVirtualItemDescReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Int64()
+  external int arg1;
+  @ffi.Uint64()
+  external int arg2;
+  @ffi.Uint64()
+  external int arg3;
 }
 
 class _RoomMessageItemTypeReturn extends ffi.Struct {
@@ -19750,6 +21385,143 @@ class _ImageDescHeightReturn extends ffi.Struct {
   external int arg1;
 }
 
+class _ImageDescThumbnailMimetypeReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Int64()
+  external int arg1;
+  @ffi.Uint64()
+  external int arg2;
+  @ffi.Uint64()
+  external int arg3;
+}
+
+class _ImageDescThumbnailSizeReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+}
+
+class _ImageDescThumbnailWidthReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+}
+
+class _ImageDescThumbnailHeightReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+}
+
+class _ImageDescThumbnailSourceReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Int64()
+  external int arg1;
+}
+
+class _VideoDescNameReturn extends ffi.Struct {
+  @ffi.Int64()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+  @ffi.Uint64()
+  external int arg2;
+}
+
+class _VideoDescMimetypeReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Int64()
+  external int arg1;
+  @ffi.Uint64()
+  external int arg2;
+  @ffi.Uint64()
+  external int arg3;
+}
+
+class _VideoDescSizeReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+}
+
+class _VideoDescWidthReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+}
+
+class _VideoDescHeightReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+}
+
+class _VideoDescBlurhashReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Int64()
+  external int arg1;
+  @ffi.Uint64()
+  external int arg2;
+  @ffi.Uint64()
+  external int arg3;
+}
+
+class _VideoDescDurationReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+}
+
+class _VideoDescThumbnailMimetypeReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Int64()
+  external int arg1;
+  @ffi.Uint64()
+  external int arg2;
+  @ffi.Uint64()
+  external int arg3;
+}
+
+class _VideoDescThumbnailSizeReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+}
+
+class _VideoDescThumbnailWidthReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+}
+
+class _VideoDescThumbnailHeightReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+}
+
+class _VideoDescThumbnailSourceReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Int64()
+  external int arg1;
+}
+
 class _FileDescNameReturn extends ffi.Struct {
   @ffi.Int64()
   external int arg0;
@@ -19774,6 +21546,45 @@ class _FileDescSizeReturn extends ffi.Struct {
   @ffi.Uint8()
   external int arg0;
   @ffi.Uint64()
+  external int arg1;
+}
+
+class _FileDescThumbnailMimetypeReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Int64()
+  external int arg1;
+  @ffi.Uint64()
+  external int arg2;
+  @ffi.Uint64()
+  external int arg3;
+}
+
+class _FileDescThumbnailSizeReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+}
+
+class _FileDescThumbnailWidthReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+}
+
+class _FileDescThumbnailHeightReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+}
+
+class _FileDescThumbnailSourceReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Int64()
   external int arg1;
 }
 
@@ -20139,6 +21950,19 @@ class _TaskListTimeZoneReturn extends ffi.Struct {
   external int arg3;
 }
 
+class _TaskListTaskBuilderReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Int64()
+  external int arg1;
+  @ffi.Uint64()
+  external int arg2;
+  @ffi.Uint64()
+  external int arg3;
+  @ffi.Int64()
+  external int arg4;
+}
+
 class _TaskListUpdateBuilderReturn extends ffi.Struct {
   @ffi.Uint8()
   external int arg0;
@@ -20150,6 +21974,15 @@ class _TaskListUpdateBuilderReturn extends ffi.Struct {
   external int arg3;
   @ffi.Int64()
   external int arg4;
+}
+
+class _GroupGetRoomIdReturn extends ffi.Struct {
+  @ffi.Int64()
+  external int arg0;
+  @ffi.Uint64()
+  external int arg1;
+  @ffi.Uint64()
+  external int arg2;
 }
 
 class _GroupTaskListDraftReturn extends ffi.Struct {
@@ -21530,6 +23363,66 @@ class _ClientTaskListsFuturePollReturn extends ffi.Struct {
   external int arg5;
 }
 
+class _ClientCreateEffektioGroupFuturePollReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint8()
+  external int arg1;
+  @ffi.Int64()
+  external int arg2;
+  @ffi.Uint64()
+  external int arg3;
+  @ffi.Uint64()
+  external int arg4;
+  @ffi.Int64()
+  external int arg5;
+}
+
+class _ClientWaitForCommentFuturePollReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint8()
+  external int arg1;
+  @ffi.Int64()
+  external int arg2;
+  @ffi.Uint64()
+  external int arg3;
+  @ffi.Uint64()
+  external int arg4;
+  @ffi.Int64()
+  external int arg5;
+}
+
+class _ClientWaitForTaskListFuturePollReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint8()
+  external int arg1;
+  @ffi.Int64()
+  external int arg2;
+  @ffi.Uint64()
+  external int arg3;
+  @ffi.Uint64()
+  external int arg4;
+  @ffi.Int64()
+  external int arg5;
+}
+
+class _ClientWaitForTaskFuturePollReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint8()
+  external int arg1;
+  @ffi.Int64()
+  external int arg2;
+  @ffi.Uint64()
+  external int arg3;
+  @ffi.Uint64()
+  external int arg4;
+  @ffi.Int64()
+  external int arg5;
+}
+
 class _UserProfileGetAvatarFuturePollReturn extends ffi.Struct {
   @ffi.Uint8()
   external int arg0;
@@ -21957,6 +23850,13 @@ class _ClientIncomingMessageRxStreamPollReturn extends ffi.Struct {
   @ffi.Uint8()
   external int arg0;
   @ffi.Int64()
+  external int arg1;
+}
+
+class _ClientSubscribeStreamPollReturn extends ffi.Struct {
+  @ffi.Uint8()
+  external int arg0;
+  @ffi.Uint8()
   external int arg1;
 }
 
