@@ -1,22 +1,22 @@
 #[cfg(test)]
 mod auth;
-#[cfg(test)]
-mod formatted_body;
-#[cfg(test)]
-mod invitation;
 // #[cfg(test)]
-// mod news;
-#[cfg(test)]
-mod reaction;
-#[cfg(test)]
-mod receipt;
-#[cfg(test)]
-mod redact;
-#[cfg(test)]
-mod reply;
-#[cfg(test)]
-mod tasks;
-#[cfg(test)]
-mod typing;
-#[cfg(test)]
-mod verification;
+// mod formatted_body;
+// #[cfg(test)]
+// mod invitation;
+// // #[cfg(test)]
+// // mod news;
+// #[cfg(test)]
+// mod reaction;
+// #[cfg(test)]
+// mod receipt;
+// #[cfg(test)]
+// mod redact;
+// #[cfg(test)]
+// mod reply;
+// #[cfg(test)]
+// mod tasks;
+// #[cfg(test)]
+// mod typing;
+// #[cfg(test)]
+// mod verification;
