@@ -15,6 +15,7 @@ mod account;
 mod auth;
 mod client;
 mod comments;
+mod common;
 mod conversation;
 mod device;
 mod group;
@@ -35,16 +36,19 @@ pub use auth::{
 };
 pub use client::{Client, ClientStateBuilder, HistoryLoadState, SyncState};
 pub use comments::{Comment, CommentDraft, CommentsManager};
+pub use common::duration_from_secs;
 pub use conversation::{Conversation, CreateConversationSettingsBuilder};
+pub use core::time::Duration as EfkDuration;
 pub use device::{DeviceChangedEvent, DeviceLeftEvent, DeviceRecord};
 pub use effektio_core::{
     events::UtcDateTime,
-    models::{Color, Faq, News, Tag},
+    models::{Color as EfkColor, Faq, News, Tag},
 };
-pub use group::{CreateGroupSettings, CreateGroupSettingsBuilder, Group};
+pub use group::{new_group_settings, CreateGroupSettings, CreateGroupSettingsBuilder, Group};
 pub use invitation::Invitation;
 pub use message::{
     FileDesc, ImageDesc, ReactionDesc, RoomEventItem, RoomMessage, RoomVirtualItem, TextDesc,
+    VideoDesc,
 };
 pub use profile::{RoomProfile, UserProfile};
 pub use receipt::{ReceiptEvent, ReceiptRecord};
@@ -58,6 +62,8 @@ pub use verification::{VerificationEmoji, VerificationEvent};
 
 #[cfg(feature = "with-mocks")]
 pub use effektio_core::mocks::*;
+
+pub use effektio_core::ruma::events::room::MediaSource;
 
 pub type DeviceId = effektio_core::ruma::OwnedDeviceId;
 pub type EventId = effektio_core::ruma::OwnedEventId;
