@@ -1,6 +1,3 @@
-use super::client::{devide_groups_from_convos, Client};
-use super::room::Room;
-use crate::api::RUNTIME;
 use anyhow::{bail, Result};
 use derive_builder::Builder;
 use effektio_core::{
@@ -14,7 +11,7 @@ use effektio_core::{
         api::client::{
             account::register::v3::Request as RegistrationRequest,
             room::{
-                create_room::v3::CreationContent, create_room::v3::Request as CreateRoomRequest,
+                create_room::v3::{CreationContent, Request as CreateRoomRequest},
                 Visibility,
             },
             uiaa,
@@ -23,7 +20,7 @@ use effektio_core::{
         events::{AnySyncTimelineEvent, MessageLikeEvent},
         room::RoomType,
         serde::Raw,
-        OwnedRoomAliasId, OwnedRoomId, OwnedUserId,
+        OwnedRoomAliasId, OwnedRoomId, OwnedUserId, UserId,
     },
     statics::{default_effektio_group_states, initial_state_for_alias},
 };
@@ -35,6 +32,13 @@ use matrix_sdk::{
     Client as MatrixClient,
 };
 use serde::{Deserialize, Serialize};
+
+use super::{
+    client::{devide_groups_from_convos, Client},
+    room::Room,
+};
+
+use crate::api::RUNTIME;
 
 #[derive(Debug, Clone)]
 pub struct Group {
@@ -145,7 +149,12 @@ impl Group {
             },
         );
     }
-    pub(crate) async fn refresh_history(&self) -> anyhow::Result<()> {
+
+    pub fn get_room_id(&self) -> String {
+        self.room_id().to_string()
+    }
+
+    pub(crate) async fn refresh_history(&self) -> Result<()> {
         let name = self.inner.name();
         tracing::trace!(name, "refreshing history");
         let room = self.inner.clone();
@@ -247,16 +256,47 @@ pub struct CreateGroupSettings {
     alias: Option<String>,
 }
 
+impl CreateGroupSettings {
+    pub fn visibility(&mut self, value: String) {
+        match value.as_str() {
+            "Public" => {
+                self.visibility = Visibility::Public;
+            }
+            "Private" => {
+                self.visibility = Visibility::Private;
+            }
+            _ => {}
+        }
+    }
+
+    pub fn add_invitee(&mut self, value: String) {
+        if let Ok(user_id) = UserId::parse(value) {
+            self.invites.push(user_id);
+        }
+    }
+
+    pub fn alias(&mut self, value: String) {
+        self.alias = Some(value);
+    }
+}
+
 // impl CreateGroupSettingsBuilder {
 //     pub fn add_invite(&mut self, user_id: OwnedUserId) {
 //         self.invites.get_or_insert_with(Vec::new).push(user_id);
 //     }
 // }
 
+pub fn new_group_settings(name: String) -> CreateGroupSettings {
+    CreateGroupSettingsBuilder::default()
+        .name(name)
+        .build()
+        .unwrap()
+}
+
 impl Client {
     pub async fn create_effektio_group(
         &self,
-        settings: CreateGroupSettings,
+        settings: Box<CreateGroupSettings>,
     ) -> Result<OwnedRoomId> {
         let c = self.client.clone();
         RUNTIME
